test form submission on mobile, refactor sign btn

+++ src/components/SignButton.jsx
@@ -1,22 +1,18 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { AiOutlineLogout } from 'react-icons/ai';
+import { AiOutlineLogout, AiOutlineLogin } from 'react-icons/ai';
 import { useContext } from 'react';
 import { UserContext } from '../context/userContext';
-const LogoutButton = () => {
+const SignButton = () => {
   const navigate = useNavigate();
-  const { setUser } = useContext(UserContext);
-  // const googleResponse = async (response) => {
-  //   await localStorage.clear();
-  //   setUser(null);
-
-  //   navigate('/login', { replace: true });
-  // };
+  const { setUser, user } = useContext(UserContext);
   const handleSignout = async (e) => {
     e.preventDefault();
-    await localStorage.clear();
-    await sessionStorage.clear();
-    setUser(null);
+    if (user) {
+      await localStorage.clear();
+      await sessionStorage.clear();
+      setUser(null);
+    }
     navigate('/login', { replace: true });
   };
   return (
@@ -24,10 +20,19 @@ const LogoutButton = () => {
       className="bg-white p-3 rounded-full flex flex-row gap-2 justify-center items-center cursor-pointer outline-none shadow-lg font-bold"
       onClick={handleSignout}
     >
-      Sign Out
-      <AiOutlineLogout color="red" fontSize={21} />
+      {!user ? (
+        <>
+          Sign In
+          <AiOutlineLogin color="red" fontSize={21} />
+        </>
+      ) : (
+        <>
+          Sign Out
+          <AiOutlineLogout color="red" fontSize={21} />
+        </>
+      )}
     </button>
   );
 };
 
-export default LogoutButton;
+export default SignButton;
